LoggedIO: start a new segment more often
Specifically, start a new segment not only if we already are over the
limit, but also if we know that the next PUT will exceed it.

IOWs, all chunks that are larger than max_segment_size will always end
up strictly in their own segments.

This is supposed to have a calculated effect: if max_segment_size is
less than minimum chunk size (2^CHUNK_MIN_EXP), metadata chunks (and
small files) will never end up in the same segments as large file
chunks.

This way, if segments are stored on a slow storage that charges for
read bandwidth, traversing archive metadata will never waste bandwidth
on downloading unrelated file chunks.